Impute missing values

Impute missing values from bootstrapped subsampling

impute_missing(E, data, nk, seed = 123456)

Arguments

E

4D array of clusterings from consensus_cluster. The number of rows is equal to the number of cases to be clustered, number of columns is equal to the clusterings obtained by different resamplings of the data, the third dimension are the different algorithms and the fourth dimension are cluster sizes.

data

data matrix with samples as rows and genes/features as columns

nk

cluster size to extract data for (single value)

seed

random seed for KNN reproducibility

Value

A list with the following elements

Details

The default output from consensus_cluster will undoubtedly contain NA entries because each replicate chooses a random subset (with replacement) of all samples. Missing values are first imputed using KNN (K-Nearest Neighbours), with the non-missing cases indicating the training set, and missing cases indicating the test set. Not all missing values are guaranteed to be imputed by KNN. See knn for details. Thus, any remaining missing values are imputed using majority voting.

Note

We consider 5 nearest neighbours and the minimum vote for definite decision is 3.

Examples

data(hgsc)
data <- t(hgsc[, -1])[1:100, 1:50]
E <- consensus_cluster(data, nk = 3:4, reps = 10, algorithms = c("hc", "km",
"sc"), progress = FALSE)
sum(is.na(E))

#> [1] 1200

E_imputed <- impute_missing(E, data, 4)
sum(is.na(E_imputed))

#> [1] 0