Consensus clustering

Runs consensus clustering across subsamples of the data, clustering algorithms, and cluster sizes.

consensus_cluster(data, nk = 2:4, pItem = 0.8, reps = 1000,
  algorithms = NULL, nmf.method = c("brunet", "lee"),
  distance = "euclidean", prep.data = c("none", "full", "sampled"),
  scale = TRUE, type = c("conventional", "robust"), min.var = 1,
  progress = TRUE, seed.nmf = 123456, seed.data = 1, save = FALSE,
  file.name = "CCOutput", time.saved = FALSE)

Arguments

data

data matrix with rows as samples and columns as variables

nk

number of clusters (k) requested; can specify a single integer or a range of integers to compute multiple k

pItem

proportion of items to be used in subsampling within an algorithm

reps

number of subsamples

algorithms

vector of clustering algorithms for performing consensus clustering. Must be any number of the following: "nmf", "hc", "diana", "km", "pam", "ap", "sc", "gmm", "block". See details. Can use a custom clustering algorithm. See example.

nmf.method

specify NMF-based algorithms to run. By default the "brunet" and "lee" algorithms are called. See nmf for details.

distance

a vector of distance functions. Defaults to "euclidean". Other options are given in dist. See example for usage of a custom distance function.

prep.data

Prepare the data on the "full" dataset, the "sampled" dataset, or "none" (default).

scale

logical; should the data be centered and scaled?

type

if we use "conventional" measures (default), then the mean and standard deviation are used for centering and scaling, respectively. If "robust" measures are specified, the median and median absolute deviation (MAD) are used.

min.var

minimum variability measure threshold used to filter the feature space for only highly variable features. Only features with a minimum variability measure across all samples greater than min.var will be used. If type = "conventional", the standard deviation is the measure used, and if type = "robust", the MAD is the measure used.

progress

logical; should a progress bar be displayed?

seed.nmf

random seed to use for NMF-based algorithms

seed.data

seed to use to ensure each algorithm operates on the same set of subsamples

save

logical; if TRUE, the returned object will be saved at each iteration as well as at the end.

file.name

file name of the written object

time.saved

logical; if TRUE, the date saved is appended to the file name. Only applicable when dir is not NULL.

Value

An array of dimension nrow(x) by reps by length(algorithms) Each slice of the array is a matrix showing consensus clustering results for algorithms. The matrices have a row for each sample, and a column for each subsample. Each entry represents a class membership.

Details

The clustering algorithms provided are:

"nmf": Nonnegative Matrix Factorization using Kullback-Leibler Divergence

"nmfEucl": Nonnegative Matrix Factorization using Euclidean distance

"hc": Hierarchical Clustering

"diana": DIvisive ANAlysis Clustering

"km": K-Means Clustering

"pam": Partition Around Mediods

"ap": Affinity Propagation

"sc": Spectral Clustering using Radial-Basis kernel function

"gmm": Gaussian Mixture Model using Bayesian Information Criterion on EM algorithm

"block": Biclustering using a latent block model

The nmf.method defaults are "brunet" (Kullback-Leibler divergence) and "lee" (Euclidean distance).

The progress bar increments for every unit of reps.

Examples

data(hgsc)
dat <- t(hgsc[, -1])

# Custom distance function
manh <- function(x) {
  stats::dist(x, method = "manhattan")
}

# Custom clustering algorithm
agnes <- function(d, k) {
  return(as.integer(stats::cutree(cluster::agnes(d, diss = TRUE), k)))
}

cc <- consensus_cluster(dat, reps = 5, algorithms = c("pam", "agnes"),
distance = c("euclidean", "manh"))

#>   |                                                                              |                                                                      |   0%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   3%  |                                                                              |====                                                                  |   5%  |                                                                              |=====                                                                 |   7%  |                                                                              |======                                                                |   8%  |                                                                              |=======                                                               |  10%  |                                                                              |========                                                              |  12%  |                                                                              |=========                                                             |  13%  |                                                                              |==========                                                            |  15%  |                                                                              |============                                                          |  17%

#> Error in get(dalgs[j]): object 'agnes' not found

str(cc)

#> List of 9
#>  $ HC_Euclidean k=2: chr [1:10] "#66C2A5" "#66C2A5" "#66C2A5" "#66C2A5" ...
#>  $ AP k=2          : chr [1:10] "#66C2A5" "#FC8D62" "#FC8D62" "#66C2A5" ...
#>  $ GMM k=2         : chr [1:10] "#66C2A5" "#FC8D62" "#FC8D62" "#FC8D62" ...
#>  $ HC_Euclidean k=3: chr [1:10] "#66C2A5" "#FC8D62" "#FC8D62" "#FC8D62" ...
#>  $ AP k=3          : chr [1:10] "#66C2A5" "#FC8D62" "#FC8D62" "#66C2A5" ...
#>  $ GMM k=3         : chr [1:10] "#66C2A5" "#FC8D62" "#FC8D62" "#FC8D62" ...
#>  $ HC_Euclidean k=4: chr [1:10] "#66C2A5" "#FC8D62" "#FC8D62" "#8DA0CB" ...
#>  $ AP k=4          : chr [1:10] "#66C2A5" "#FC8D62" "#8DA0CB" "#66C2A5" ...
#>  $ GMM k=4         : chr [1:10] "#66C2A5" "#FC8D62" "#8DA0CB" "#FC8D62" ...