Combine, evaluate, trim, and reweigh algorithms

consensus_combine combines results for multiple objects from consensus_cluster() and outputs either the consensus matrices or consensus classes for all algorithms. consensus_evaluate evaluates algorithms on internal/external validation indices. consensus_weigh weighs clustering algorithms based on these two indices. consensus_trim removes algorithms that rank low on internal indices before using in consensus functions.

consensus_combine(..., element = c("matrix", "class"))

consensus_evaluate(data, ..., cons.cl = NULL, ref.cl = NULL, plot = FALSE)

consensus_trim(data, ..., cons.cl = NULL, ref.cl = NULL, quantile = 0.75,
  reweigh = FALSE, show.eval = TRUE)

Arguments

...
any number of objects outputted from consensus_cluster

element
either "matrix" or "class" to extract the consensus matrix or consensus class, respectively.

data
data matrix with rows as samples and columns as variables

cons.cl
matrix of cluster assignments from consensus functions such as kmodes and majority_voting

ref.cl
reference class

plot
logical; if TRUE, graph_all is called

quantile
if an algorithm's sum of ranks across indices is in the quantile quantile, the algorithm is kept. Otherwise it is removed (trimmed).

reweigh
logical; if TRUE, after trimming out poor performing algorithms, each algorithm is reweighed depending on its internal indices.

show.eval
logical; if TRUE (default), show the evaluation output from consensus_evaluate

Value

consensus_combine returns either a list of all consensus matrices or a data frame showing all the consensus classes consensus_evaluate returns a list with the following elements

k

if ref.cl is not NULL, this is the number of distinct classes in the reference; otherwise the chosen k is determined by the one giving the largest mean PAC across algorithms

pac

a data frame showing the PAC for each combination of algorithm and cluster size

internal

a list of data frames for all k showing internal evaluation indices

external

a data frame showing external evaluation indices for k

consensus_trim returns a list with three elements
alg.keep

algorithms kept

alg.remove

algorithms removed

eval

if show.eval = TRUE, the evaluation output is returned, otherwise NULL

data.new

A new version of a consensus_cluster data object. Potentially no trimming depending on quantile

value.
Details

consensus_combine is useful for generating summaries because the results have been combined into a single object. For example, if element = "class", then the resulting object can be used to create a consensus matrix across algorithms, which can be visualized as a heatmap. consensus_evaluate always shows internal indices. If ref.cl is not NULL, external indices are shown in addition to internal indices. Relevant graphical displays are also outputted. consensus_trim ranks algorithms by internal indices used in consensus_evaluate. The sum of the ranks for each algorithm is used as the measure of comparison. This also means the magnitude of the internal indices is not taken into account.

Examples

# Consensus clustering for multiple algorithms
set.seed(911)
x <- matrix(rnorm(500), ncol = 10)
CC1 <- consensus_cluster(x, nk = 3:4, reps = 10, algorithms = "ap",
progress = FALSE)
CC2 <- consensus_cluster(x, nk = 3:4, reps = 10, algorithms = "gmm",
progress = FALSE)

# Combine and return either matrices or classes
y1 <- consensus_combine(CC1, CC2, element = "matrix")
str(y1)

#> List of 2
#>  $ 3:List of 2
#>   ..$ AP : num [1:50, 1:50] 1 0.429 0.375 0.222 0.714 ...
#>   ..$ GMM: num [1:50, 1:50] 1 0 0.333 0 1 ...
#>  $ 4:List of 2
#>   ..$ AP : num [1:50, 1:50] 1 0.143 0 0 0.857 ...
#>   ..$ GMM: num [1:50, 1:50] 1 0 0.333 0 0.5 ...

y2 <- consensus_combine(CC1, CC2, element = "class")
str(y2)

#> List of 2
#>  $ 3: int [1:50, 1:2] 1 2 1 2 1 2 1 1 3 1 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "AP" "GMM"
#>  $ 4: int [1:50, 1:2] 1 2 3 2 1 2 3 2 4 3 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "AP" "GMM"

# Evaluate algorithms on internal and external indices and make plots
set.seed(1)
ref.cl <- sample(1:4, 50, replace = TRUE)
z <- consensus_evaluate(x, CC1, CC2, ref.cl = ref.cl, plot = FALSE)

# Trim algorithms: remove those that rank low on internal indices
CC3 <- consensus_trim(x, CC1, CC2, ref.cl = ref.cl, quantile = 0.8)
str(CC3, max.level = 2)

#> List of 4
#>  $ alg.keep  : chr [1:2] "AP" "GMM"
#>  $ alg.remove: chr(0) 
#>  $ eval      :List of 4
#>   ..$ k       : int 4
#>   ..$ pac     :'data.frame':	2 obs. of  3 variables:
#>   ..$ internal:List of 2
#>   ..$ external:'data.frame':	2 obs. of  19 variables:
#>  $ data.new  : num [1:50, 1:10, 1:2, 1] 2 2 1 3 2 NA 4 1 4 3 ...
#>   ..- attr(*, "dimnames")=List of 4