Diverse Clustering Ensemble

Runs consensus clustering across subsamples, algorithms, and number of clusters (k).

dice(data, nk, reps = 10, algorithms = NULL, nmf.method = c("brunet",
  "lee"), distance = "euclidean", cons.funs = c("kmodes", "CSPA",
  "majority", "LCE"), sim.mat = c("cts", "srs", "asrs"),
  prep.data = c("none", "full", "sampled"), min.var = 1, seed = 1,
  trim = FALSE, reweigh = FALSE, evaluate = TRUE, plot = FALSE,
  ref.cl = NULL, progress = TRUE)

Arguments

data

matrix with rows as observations, columns as variables

nk

number of clusters (k) requested; can specify a single integer or a range of integers to compute multiple k

reps

number of data subsamples to generate. See consensus_cluster for details.

algorithms

clustering algorithms to be used in the ensemble. Current options are "nmf", "hc", "diana", "km", "pam", "ap", "sc", "gmm", "block". See consensus_cluster for details.

nmf.method

specify NMF-based algorithms to run. By default the "brunet" and "lee" algorithms are called. See consensus_cluster for details.

distance

a vector of distance functions. Defaults to "euclidean". Can use a custom distance function. See consensus_cluster for details.

cons.funs

consensus functions to use. Current options are "kmodes" (k-modes), "majority" (majority voting), "CSPA" (Cluster-based Similarity Partitioning Algorithm), "LCE" (linkage clustering ensemble)

sim.mat

type of similarity matrix. One of "cts", "srs", "asrs. See LCE for details.

prep.data

Prepare the data on the "full" dataset, the "sampled" dataset, or "none" (default). See Details.

min.var

minimum variability measure threshold. See prepare_data.

seed

seed used for imputation

trim

logical; if TRUE, the number of algorithms in algorithms is reduced based on internal validity index performance prior to consensus clustering by cons.funs. Defaults to FALSE.

reweigh

logical; if TRUE, algorithms are reweighted based on internal validity index performance after trimming. Well-performing algorithms are given higher weight prior to consensus clustering by cons.funs. Defaults to FALSE. Ignored if trim = FALSE.

evaluate

logical; if TRUE (default), validity indices are returned. Internal validity indices are always computed. If ref.cl is not NULL, then external validity indices will also be computed.

plot

logical; if TRUE, graph_all is called and relevant graphs are outputted. Ignored if evaluate = FALSE.

ref.cl

reference class; a vector of length equal to the number of observations.

progress

logical; if TRUE (default), progress bar is shown.

Value

A final clustering assignment from the diverse clustering ensemble method.

Details

There are three ways to handle the input data before clustering via argument prep.data. The default is to use the raw data as-is ("none"). Or, we can enact prepare_data on the full dataset ("full"), or the bootstrap sampled datasets ("sampled").

Examples

library(dplyr)
data(hgsc)
dat <- t(hgsc[, -1])
ref.cl <- data.frame(initCol = rownames(dat)) %>%
tidyr::separate(initCol,
                into = c("patientID", "Class"),
                sep = "_") %>%
  magrittr::use_series(Class) %>%
  factor() %>%
  as.integer()
dice.obj <- dice(dat, nk = 4, reps = 3, algorithms = "hc", cons.funs =
"kmodes", ref.cl = ref.cl)

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======================                                               |  33%  |                                                                              |===============================================                       |  67%  |                                                                              |======================================================================| 100%

str(dice.obj, max.level = 2)

#> List of 2
#>  $ clusters: int [1:489, 1:2] 4 3 1 3 3 4 1 3 2 4 ...
#>   ..- attr(*, "dimnames")=List of 2
#>  $ indices :List of 4
#>   ..$ k       : int 4
#>   ..$ pac     :'data.frame':	1 obs. of  2 variables:
#>   ..$ internal:List of 1
#>   ..$ external:'data.frame':	2 obs. of  19 variables: