diceR: Diverse Cluster Ensemble in R

Performs cluster analysis using an ensemble clustering framework. Results from a diverse set of algorithms are pooled together using methods such as majority voting, K-Modes, LinkCluE, and CSPA. There are options to compare cluster assignments across algorithms using internal and external indices, visualizations such as heatmaps, and significance testing for the existence of clusters.

See also

Useful links:

https://github.com/AlineTalhouk/diceR

https://alinetalhouk.github.io/diceR

Report bugs at https://github.com/AlineTalhouk/diceR/issues