External validity indices

External validity indices compare a predicted clustering result with a reference class or gold standard.

ev_nmi(pred.lab, ref.lab, method = "emp")

ev_confmat(pred.lab, ref.lab)

Arguments

pred.lab
predicted labels generated by classifier

ref.lab
reference labels for the observations

method
method of computing the entropy. Can be any one of "emp", "mm", "shrink", or "sg".

Value

ev_nmi returns the normalized mutual information.

ev_confmat returns a vector of the following metrics: overall accuracy, Cohen's kappa, no information rate, accuracy p-value. Statistics are difficult to compare when there are multiclass comparisons. We hence also report averaged statistics for: sensitivity, specificity, PPV, NPV, prevalence, detection rate, detection prevalence, accuracy, and balanced accuracy.

Details

ev_nmi calculates the normalized mutual information

ev_confmat calculates a variety of statistics associated with confusion matrices.

Note

ev_nmi is adapted from infotheo::mutinformation()

References

Strehl A, Ghosh J. Cluster ensembles: a knowledge reuse framework for combining multiple partitions. J. Mach. Learn. Res. 2002;3:583-617.

Examples

suppressWarnings(RNGversion("3.5.0"))
set.seed(1)
E <- matrix(rep(sample(1:4, 1000, replace = TRUE)), nrow = 100, byrow =
              FALSE)
x <- sample(1:4, 100, replace = TRUE)
y <- sample(1:4, 100, replace = TRUE)
ev_nmi(x, y)

#> [1] 0.02372827

ev_confmat(x, y)

#>          Overall Accuracy             Cohen's kappa       No Information Rate 
#>                0.31000000                0.07716999                0.32000000 
#>       P-Value [Acc > NIR]       Average Sensitivity       Average Specificity 
#>                0.62159446                0.30251157                0.77010015 
#>               Average PPV               Average NPV    Average Detection Rate 
#>                0.30685185                0.76935412                0.07750000 
#>          Average Accuracy Average Balanced Accuracy 
#>                0.65500000                0.53630586