Minimize Frobenius norm for between two matrices

Finds a permutation of a matrix such that its Frobenius norm with another matrix is minimized.

min_fnorm(A, B = diag(nrow(A)))

Arguments

A
data matrix we want to permute

B
matrix whose distance with the permuted A we want to minimize. By default, B <- diag(nrow(A)), so the permutation maximizes the trace of A.

Value

Permuted matrix such that it is the permutation of A closest to B

Details

Finds the permutation P of A such that ||PA - B|| is minimum in Frobenius norm. Uses the linear-sum assignment problem (LSAP) solver in the package clue. The default B is the identity matrix of same dimension, so that the permutation of A maximizes its trace. This procedure is useful for constructing a confusion matrix when we don't know the true class labels of a predicted class and want to compare to a reference class.

Examples

set.seed(1)
A <- matrix(sample(1:25, size = 25, rep = FALSE), 5, 5)
min_fnorm(A)

#> $pmat
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]   20   11   21   23   15
#> [2,]    7   18    4   19    6
#> [3,]   14   12   24    8    2
#> [4,]    9   22    3   25   16
#> [5,]    5    1   13   10   17
#> 
#> $perm
#> Optimal assignment:
#> 1 => 4, 2 => 1, 3 => 3, 4 => 2, 5 => 5
#> 
#> $ord
#> [1] 2 4 3 1 5
#>